refactor: move exceptions to model package and handler to its own package
