+++ src/main/java/gorbiel/stock_sim/exception/model/BadRequestException.java
@@ -1,4 +1,4 @@
-package gorbiel.stock_sim.exception;
+package gorbiel.stock_sim.exception.model;
 
 public class BadRequestException extends RuntimeException {
