+++ src/main/java/gorbiel/stock_sim/exception/model/InsufficientBankStockException.java
@@ -1,4 +1,4 @@
-package gorbiel.stock_sim.exception;
+package gorbiel.stock_sim.exception.model;
 
 public class InsufficientBankStockException extends BadRequestException {
