+++ src/main/java/gorbiel/stock_sim/wallet/service/WalletStockOperationServiceImpl.java
@@ -5,8 +5,8 @@ import gorbiel.stock_sim.audit.model.OperationType;
 import gorbiel.stock_sim.audit.repository.AuditLogEntryRepository;
 import gorbiel.stock_sim.bank.model.BankStockHolding;
 import gorbiel.stock_sim.bank.service.BankStockService;
-import gorbiel.stock_sim.exception.InsufficientBankStockException;
-import gorbiel.stock_sim.exception.InsufficientWalletStockException;
+import gorbiel.stock_sim.exception.model.InsufficientBankStockException;
+import gorbiel.stock_sim.exception.model.InsufficientWalletStockException;
 import gorbiel.stock_sim.wallet.model.Wallet;
 import gorbiel.stock_sim.wallet.model.WalletStockHolding;
 import gorbiel.stock_sim.wallet.repository.WalletRepository;
