+++ src/test/java/gorbiel/stock_sim/wallet/service/WalletStockOperationServiceImplTest.java
@@ -7,8 +7,8 @@ import gorbiel.stock_sim.audit.model.OperationType;
 import gorbiel.stock_sim.audit.repository.AuditLogEntryRepository;
 import gorbiel.stock_sim.bank.model.BankStockHolding;
 import gorbiel.stock_sim.bank.repository.BankStockHoldingRepository;
-import gorbiel.stock_sim.exception.BadRequestException;
-import gorbiel.stock_sim.exception.ResourceNotFoundException;
+import gorbiel.stock_sim.exception.model.BadRequestException;
+import gorbiel.stock_sim.exception.model.ResourceNotFoundException;
 import gorbiel.stock_sim.wallet.model.Wallet;
 import gorbiel.stock_sim.wallet.model.WalletStockHolding;
 import gorbiel.stock_sim.wallet.repository.WalletRepository;
